Change package name, and move PPX executable
Three things in this commit! I initially tried to split
it into couple of commits, but changes ended up
overlapping so I threw them all into one PR.

- Change the package name to `decco`
- Move the ppx executable to the root level
  and called it `ppx`.  (see reasoning below)
- Change the ppx executable to use bash
  instead of Node for perffffffffffffffff.

Notes on the moving of the PPX:

The current state of bucklescript ppxs is rather confusing,
each having its own executable path. Every time the user
wants to install a PPX, they have to check the readme to see
what particular location they have to point to for the
executable. With `bs-let` I moved the executable to the
root level and named it `ppx`, so that when the user is
installing, all they have to do is add `package_name/ppx` to
the ppx flags to get it installed and working. If all bucklescript
PPXs followed this pattern, users wouldn't need to refer to the
readme at all after the first few times.

What do you think? @ryb73